requirements.txt
albumentations==0.4.6
catalyst @ git+https://github.com/catalyst-team/catalyst@f5aeaccdde288428ce53d6950e73540cc6d9874a
jupyter==1.0.0
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.1.2.30
opencv-python-headless==4.4.0.44
pandas==1.1.2
Pillow==7.2.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
segmentation-models-pytorch==0.1.2
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
wandb==0.10.3
